# Conflicts:
#	ingredient_service/src/main/java/com/recipe/ingredient_service/service/IngredientService.java
#	recipe_service/src/main/java/com/recipe/recipe_service/controller/RecipeController.java

+++ recipe_service/src/main/java/com/recipe/recipe_service/repository/RecipeLikesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RecipeLikesRepository extends JpaRepository<RecipeLikes, Integer> {
+public interface RecipeLikesRepository extends JpaRepository<RecipeLikes, Long> {
     // 특정 유저와 레시피에 대해 좋아요가 존재하는지 확인
     Optional<RecipeLikes> findByRecipeIdAndUserId(Long recipeId, Long userId);
